Removed unused code, fix warnings, code path tweaking
