+++ Cargo.toml
@@ -21,6 +21,8 @@ default = ["std"]
 std = []
 # SMT implemented in C
 smtc = []
+# A storage optimized SMT implemented in trie (https://ouvrard-pierre-alain.medium.com/sparse-merkle-tree-86e6e2fc26da)
+trie = []
 
 [dependencies]
 cfg-if = "0.1"
@@ -33,7 +35,7 @@ rand = "0.8"
 hex = "0.4.3"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-anyhow = "1.0"
+anyhow = "1.0.65"
 
 [[bench]]
 name = "smt_benchmark"
