+++ src/lib.rs
@@ -62,7 +62,6 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 pub mod blake2b;
-pub mod branch;
 #[cfg(feature = "smtc")]
 pub mod ckb_smt;
 pub mod default_store;
@@ -80,7 +79,10 @@ pub mod trie_tree;
 pub use ckb_smt::{SMTBuilder, SMT};
 pub use h256::H256;
 pub use merkle_proof::{CompiledMerkleProof, MerkleProof};
-//pub use tree::SparseMerkleTree;
+#[cfg(not(feature = "trie"))]
+pub use tree::SparseMerkleTree;
+pub use tree::{BranchKey, BranchNode};
+#[cfg(feature = "trie")]
 pub use trie_tree::SparseMerkleTree;
 
 /// Expected path size: log2(256) * 2, used for hint vector capacity
