+++ src/merge.rs
@@ -85,6 +85,8 @@ impl MergeValue {
         }
     }
 
+    /// Helper function for get base_node
+    /// When call with MergeValue::Value(v), it would be v
     pub fn base_node<H: Hasher + Default>(&self) -> H256 {
         match self {
             MergeValue::ShortCut {
@@ -104,6 +106,8 @@ impl MergeValue {
         }
     }
 
+    /// Helper function for Shortcut node
+    /// Transform it into a MergeWithZero node
     pub fn into_merge_with_zero<H: Hasher + Default>(&self) -> MergeValue {
         match self {
             MergeValue::ShortCut { key, value, height } => {
@@ -121,12 +125,7 @@ impl MergeValue {
                     zero_count: *height,
                 }
             }
-            MergeValue::MergeWithZero {
-                base_node: _,
-                zero_bits: _,
-                zero_count: _,
-            } => self.clone(),
-            MergeValue::Value(_) => {
+            _ => {
                 unreachable!();
             }
         }
