+++ src/trie_tree.rs
@@ -1,17 +1,13 @@
-use crate::merge::hash_base_node;
-use crate::merge::MergeValue::MergeWithZero;
 use crate::{
-    branch::*,
     error::{Error, Result},
-    merge,
     merge::{merge, MergeValue},
     merkle_proof::MerkleProof,
     traits::{Hasher, StoreReadOps, StoreWriteOps, Value},
+    tree::{BranchKey, BranchNode},
     vec::Vec,
     H256, MAX_STACK_SIZE,
 };
 use core::marker::PhantomData;
-use std::thread::current;
 
 /// Sparse merkle tree
 #[derive(Default, Debug)]
@@ -21,7 +17,7 @@ pub struct SparseMerkleTree<H, V, S> {
     phantom: PhantomData<(H, V)>,
 }
 
-impl<H, V, S: StoreReadOps<V>> SparseMerkleTree<H, V, S> {
+impl<H, V, S> SparseMerkleTree<H, V, S> {
     /// Build a merkle tree from root and store
     pub fn new(root: H256, store: S) -> SparseMerkleTree<H, V, S> {
         SparseMerkleTree {
@@ -57,7 +53,7 @@ impl<H, V, S: StoreReadOps<V>> SparseMerkleTree<H, V, S> {
     }
 }
 
-impl<H: Hasher + Default, V: Value + PartialEq, S: StoreReadOps<V> + StoreWriteOps<V>>
+impl<H: Hasher + Default, V: Value, S: StoreReadOps<V> + StoreWriteOps<V>>
     SparseMerkleTree<H, V, S>
 {
     /// Update a leaf, return new merkle root
@@ -73,7 +69,7 @@ impl<H: Hasher + Default, V: Value + PartialEq, S: StoreReadOps<V> + StoreWriteO
         }
 
         let mut last_height = core::u8::MAX;
-        while last_height >= 0 {
+        loop {
             // walk from top to bottom
             let node_key = key.parent_path(last_height);
             let branch_key = BranchKey::new(last_height, node_key); // this represents a position in the tree
@@ -308,11 +304,11 @@ impl<H: Hasher + Default, V: Value, S: StoreReadOps<V>> SparseMerkleTree<H, V, S
                     };
 
                     match target {
-                        MergeValue::ShortCut { key, height: h, .. } => {
+                        MergeValue::ShortCut { key, .. } => {
                             if !sibling.is_zero() {
                                 bitmap.set_bit(height);
                             }
-                            if !key.eq(current_key){
+                            if !key.eq(current_key) {
                                 let fork_height = key.fork_height(current_key);
 
                                 bitmap.set_bit(fork_height);
@@ -355,7 +351,6 @@ impl<H: Hasher + Default, V: Value, S: StoreReadOps<V>> SparseMerkleTree<H, V, S
             }
 
             let mut proof_result = Vec::new();
-            let mut inserted_index = Vec::new();
             for height in (0..=fork_height).rev() {
                 if height == fork_height && leaf_index + 1 < keys.len() {
                     // If it's not final round, we don't need to merge to root (height=255)
@@ -376,19 +371,20 @@ impl<H: Hasher + Default, V: Value, S: StoreReadOps<V>> SparseMerkleTree<H, V, S
                         };
 
                         match current {
-                            MergeValue::ShortCut { key, value, height } => {
+                            MergeValue::ShortCut { key, value, .. } => {
                                 if !sibling.is_zero() {
-                                    if sibling.is_shortcut() {
-                                        proof_result.push(sibling.into_merge_with_zero::<H>());
-                                    } else {
-                                        proof_result.push(sibling);
-                                    }
+                                    push_result_maybe_shortcut::<H>(&mut proof_result, sibling);
                                 }
                                 if !key.eq(&leaf_key) {
-                                    // this means key does not exist // FIXME!
+                                    // this means key does not exist
                                     let fork_height = key.fork_height(&leaf_key);
-                                    if leaves_bitmap[leaf_index].get_bit(fork_height) && !skip_height.contains(&fork_height) {
-                                        proof_result.push(MergeValue::shortcut(key, value, fork_height).into_merge_with_zero::<H>());
+                                    if leaves_bitmap[leaf_index].get_bit(fork_height)
+                                        && !skip_height.contains(&fork_height)
+                                    {
+                                        proof_result.push(
+                                            MergeValue::shortcut(key, value, fork_height)
+                                                .into_merge_with_zero::<H>(),
+                                        );
                                     }
 
                                     if fork_height == 1 && leaves_bitmap[leaf_index].get_bit(0) {
@@ -397,58 +393,51 @@ impl<H: Hasher + Default, V: Value, S: StoreReadOps<V>> SparseMerkleTree<H, V, S
                                 }
                                 break;
                             }
-                            MergeValue::MergeWithZero { zero_count, .. } => {
+                            _ => {
                                 if !sibling.is_zero() {
-                                    if sibling.is_shortcut() {
-                                        proof_result.push(sibling.into_merge_with_zero::<H>());
-                                    } else {
-                                        proof_result.push(sibling);
-                                    }
-                                }
-                            }
-                            MergeValue::Value(_) => {
-                                if !sibling.is_zero() {
-                                    match sibling {
-                                        MergeValue::ShortCut { .. } => {
-                                            proof_result.push(sibling.into_merge_with_zero::<H>());
-                                        }
-                                        _ => {
-                                            proof_result.push(sibling);
-                                        }
-                                    }
+                                    push_result_maybe_shortcut::<H>(&mut proof_result, sibling);
                                 }
                             }
                         }
-                    }  else if !inserted_index.contains(&height) {
-                        // find from up to down
+                    } else {
+                        // Maybe we've skipped shortcut node, find from up to down
                         for i in (height..=core::u8::MAX).rev() {
                             let parent_key = leaf_key.parent_path(i);
                             let is_right = leaf_key.is_right(i);
                             let parent_branch_key = BranchKey::new(i, parent_key);
-                            if let Some(parent_branch) = self.store.get_branch(&parent_branch_key)? {
-                                let (sibling, current) = if is_right {
-                                    (parent_branch.left, parent_branch.right)
+                            if let Some(parent_branch) =
+                                self.store.get_branch(&parent_branch_key)?
+                            {
+                                let current = if is_right {
+                                    parent_branch.right
                                 } else {
-                                    (parent_branch.right, parent_branch.left)
+                                    parent_branch.left
                                 };
 
                                 match current {
-                                    MergeValue::ShortCut { key, value, height:h} => {
+                                    MergeValue::ShortCut {
+                                        key,
+                                        value,
+                                        ..
+                                    } => {
                                         if !key.eq(&leaf_key) {
                                             let fork_at = key.fork_height(&leaf_key);
                                             if fork_at == height {
-                                                let insert_value = MergeValue::shortcut(key, value, height).into_merge_with_zero::<H>();
-                                                proof_result.push(insert_value);
+                                                proof_result.push(
+                                                    MergeValue::shortcut(key, value, height)
+                                                        .into_merge_with_zero::<H>(),
+                                                );
                                             }
                                         }
-                                    },
+                                        break;
+                                    }
                                     _ => {
                                         continue;
                                     }
                                 }
-
                             }
                         }
+                        break; // we should stop further looping
                     }
                 }
             }
@@ -461,5 +450,14 @@ impl<H: Hasher + Default, V: Value, S: StoreReadOps<V>> SparseMerkleTree<H, V, S
             leaf_index += 1;
         }
         Ok(MerkleProof::new(leaves_bitmap, proof))
+    }
+}
+
+/// Helper function for a merkle_path insertion
+fn push_result_maybe_shortcut<H: Hasher + Default>(proof_result: &mut Vec<MergeValue>, value: MergeValue) {
+    if value.is_shortcut(){
+        proof_result.push(value.into_merge_with_zero::<H>())
+    } else {
+        proof_result.push(value)
     }
 }
